// src/interfaces/pokemonInterfaces.tsx
export interface SimplePokemon {
    id:         string;
    name:       string;
    picture:    string;
    color?:     string;
}